requirements.txt
# Core runtime
fastapi
uvicorn
apscheduler
sqlalchemy
pydantic
python-dotenv

# CLI / diagnostics
rich

# Data helpers
pyyaml
yt-dlp
